Fix test application context
